backproj/requirements.txt
django
django-ninja
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
python-dotenv
pika
